refactor: rename file to ReportModalContent
